Update run-container to include short example for config (#2886)
Added a short example to explain what should actually be set as a OPENVPN_CONFIG variable. 

Technical deep dive is great and interesting, but should come later. The most needed basic information to use the parameter should come first, further explanation should come afterwards.

Bit of background/explanation: I was so confused by what to put in the CONFIG variable that I messed around with a manual ovpn instead. Judging by #2408 I'm not the only one. I realize the same explanation as in the issue is present in this wiki page, but it's far down, and opening the github page is much quicker and easier imo. (If you're in the docs you already have a browser open, but one might configure the container through portainer instead of a terminal/ssh connection)